+++ apps/www/src/app/(main)/(docs)/components/charts/area-chart/examples/area-chart-1.tsx
@@ -2,13 +2,12 @@
 
 "use client";
 
-import { Card, CardContent, CardHeader } from "@jamsr-ui/react";
+import {  Card, CardContent, CardHeader } from "@jamsr-ui/react";
 import {
   Area,
   AreaChart,
   CartesianGrid,
   ResponsiveContainer,
-  Tooltip,
   XAxis,
   YAxis,
 } from "recharts";
@@ -73,43 +72,6 @@ export const AreaChart1 = () => {
                 style: {
                   fontSize: "var(--fs-xs)",
                 },
-              }}
-            />
-            <Tooltip
-              cursor={false}
-              content={({ payload, label }) => {
-                return (
-                  <div className="flex h-auto min-w-[120px] flex-col items-center gap-x-2 rounded-md bg-background p-2 text-xs shadow-sm">
-                    <div className="flex w-full flex-col gap-y-1">
-                      <span className="font-medium text-foreground">
-                        {label}
-                      </span>
-                      {payload?.map((item, idx) => {
-                        const { value } = item;
-                        const { name } = item;
-                        const category =
-                          categories.find((c) => c.toLowerCase() === name) ??
-                          name;
-                        return (
-                          <div
-                            key={idx}
-                            className="flex w-full items-center gap-x-2"
-                          >
-                            <div
-                              className="size-2 flex-none rounded-full"
-                              style={{
-                                backgroundColor: `hsl(var(--ui-success-${(idx + 1) * 200}))`,
-                              }}
-                            />
-                            <span className="text-default-500">
-                              {value} {category}
-                            </span>
-                          </div>
-                        );
-                      })}
-                    </div>
-                  </div>
-                );
               }}
             />
             <defs>