+++ apps/www/src/app/(main)/(docs)/components/charts/bar-chart/page.tsx
@@ -1,5 +1,6 @@
 import { VariantPage } from "@/components/docs/variant-page";
 import { VariantWrapper } from "@/components/docs/variant-wrapper";
+import { type Metadata } from "next";
 import { BarChartMultiColumns } from "./examples/bar-chart-multi-columns";
 import { BarChartOneColumn } from "./examples/bar-chart-one-column";
 import { BarChartStack } from "./examples/bar-chart-stack";
@@ -8,29 +9,33 @@ import { BarChartVerticalMultiColumns } from "./examples/bar-chart-vertical-mult
 import { BarChartVerticalStack } from "./examples/bar-chart-vertical-stack";
 
 const title = "Bar Chart";
-export const metadata = {
+const description = `A bar chart visually represents data using rectangular bars, 
+where the length of each bar corresponds to its value.`;
+
+export const metadata: Metadata = {
   title,
+  description,
 };
 
 const Page = () => {
   return (
-    <VariantPage heading={title}>
-      <VariantWrapper heading="One Column">
+    <VariantPage heading={title} description={description}>
+      <VariantWrapper heading="Single Column">
         <BarChartOneColumn />
       </VariantWrapper>
-      <VariantWrapper heading="Multi Columns">
+      <VariantWrapper heading="Multiple Column">
         <BarChartMultiColumns />
       </VariantWrapper>
-      <VariantWrapper heading="Stack">
+      <VariantWrapper heading="Stacked Column">
         <BarChartStack />
       </VariantWrapper>
-      <VariantWrapper heading="Vertical One Column">
+      <VariantWrapper heading="Vertical Single Column">
         <BarChartVerticalOneColumn />
       </VariantWrapper>
-      <VariantWrapper heading="Vertical Multi Columns">
+      <VariantWrapper heading="Vertical Multiple Columns">
         <BarChartVerticalMultiColumns />
       </VariantWrapper>
-      <VariantWrapper heading="Vertical Stack">
+      <VariantWrapper heading="Vertical Stacked Column">
         <BarChartVerticalStack />
       </VariantWrapper>
     </VariantPage>