+++ apps/www/src/app/(main)/(docs)/components/charts/pie-chart/page.tsx
@@ -1,14 +1,19 @@
 import { VariantPage } from "@/components/docs/variant-page";
 import { PieChart1 } from "./examples/pie-chart-1";
+import { type Metadata } from "next";
 
 const title = "Pie Chart";
-export const metadata = {
+const description = `A pie chart represents data as slices of a circular pie, 
+showing proportions and percentages of a whole.`;
+
+export const metadata: Metadata = {
   title,
+  description,
 };
 
 const Page = () => {
   return (
-    <VariantPage heading={title}>
+    <VariantPage heading={title} description={description}>
       <PieChart1 />
     </VariantPage>
   );