+++ packages/components/chart/src/core/line.tsx
@@ -1,0 +1,6 @@
+import { type ComponentProps } from "react";
+import { Line as LineBase } from "recharts";
+
+export const Line = (props: ComponentProps<typeof LineBase>) => {
+  return <LineBase {...props} />;
+};