+++ packages/components/chart/src/core/pie.tsx
@@ -1,0 +1,6 @@
+import { type ComponentProps } from "react";
+import { Pie as PieBase } from "recharts";
+
+export const Pie = (props: ComponentProps<typeof PieBase>) => {
+  return <PieBase {...props} />;
+};