+++ packages/components/chart/src/index.ts
@@ -1,0 +1,14 @@
+"use client";
+
+export { ChartTooltip } from "./chart-tooltip";
+export { AreaChart } from "./core/area-chart";
+export { ChartContainer } from "./chart-container";
+export * from "./types";
+
+export { Area } from "./core/area";
+export { Bar } from "./core/bar";
+export { CartesianGrid } from "./core/cartesian-grid";
+export { Line } from "./core/line";
+export { Pie } from "./core/pie";
+export { XAxis } from "./core/x-axis";
+export { YAxis } from "./core/y-axis";