+++ packages/components/chart/src/types.ts
@@ -1,0 +1,7 @@
+export type ChartConfig = {
+  [k in string]: {
+    label?: React.ReactNode;
+    icon?: React.ComponentType;
+    color?: string;
+  };
+};